TileEditor (should be renamed to TileSetEditor) now has an extra grid for showing tile images from the selected directory.
Tile graphics can be updated through the new grid, but changes aren't saved yet.
Grid got an extra property: offset. This can be used to indicate where on the window the grid is placed. Not sure though if this should be a Grid property. It is used to draw at an offset of the window.
Grid#place_tile! has been added, which does what the force=true setting of Grid#place_tile used to do.